api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
